secp_omni: Rename keypair to secret_key
