+++ internal/secp256k1_omni_context/src/lib.rs
@@ -1,7 +1,7 @@
 #![warn(unused_extern_crates, missing_debug_implementations, rust_2018_idioms)]
 #![forbid(unsafe_code)]
 
-pub use crate::keypair::*;
+pub use crate::secret_key::*;
 pub use secp256k1;
 
-mod keypair;
+mod secret_key;
