+++ internal/secp256k1_omni_context/src/secret_key.rs
